+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/src/Audacia.CodeAnalysis.Analyzers/Common/DiagnosticId.cs
@@ -19,6 +19,6 @@
         public const string DoNotUseNumberInIdentifierName = "ACL1014";
         public const string DoNotUseProducesResponseTypeWithTypedResults = "ACL1015";
         public const string UseTypedResultsInsteadOfIActionResult = "ACL1016";
-        public const string SupressionMustHaveJustification = "ACL1017";
+        public const string SupressionRequiresJustification = "ACL1018";
     }
 }
