+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/src/Audacia.CodeAnalysis.Analyzers/Common/HelpLinkUrlFactory.cs
@@ -56,7 +56,7 @@ namespace Audacia.CodeAnalysis.Analyzers.Common
                     return $"{helpLinkUrlPrefix}controller-action-has-producesresponsetype-attribute-when-return-type-is-typedresults";
                 case DiagnosticId.UseTypedResultsInsteadOfIActionResult:
                     return $"{helpLinkUrlPrefix}controller-action-should-return-typedresults-instead-of-iactionresults";                
-                case DiagnosticId.SupressionMustHaveJustification:
+                case DiagnosticId.SupressionRequiresJustification:
                     return $"{helpLinkUrlPrefix}code-analysis-supression-attribute-requires-justification";
                 default:
                     throw new ArgumentOutOfRangeException(
